+++ NEWS.md
@@ -16,6 +16,12 @@
 
 * The `$model` of a `LearnerTorch` now has a printer.
 * Added more image learners from {torchvision}.
+* `LearnerTorch` now tracks the validation scores of the best epoch and exposes them via the new
+  `$best_valid_scores` field, so `msr("best_valid_score")` can be used for tuning.
+  This is the epoch that is also reported via `$internal_tuned_values`, whereas `$internal_valid_scores`
+  refers to the epoch of the network that is actually returned, i.e. to the last epoch unless
+  `restore_best_weights` is `TRUE`.
+  Tracking requires early stopping to be active (`patience > 0`).
 * Most `LearnerTorchVision` are now `jittable`.
 * Ported the `TabM` tabular learner from Python.
 * `LearnerTorch` now has `.loss_fn(task, param_vals)` private method that allows